Move the `testing-common` module under the `@bayou` namespace.
