+++ local-modules/@bayou/testing-common/BaseReporter.js
