+++ local-modules/@bayou/testing-common/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "testing-common",
+  "name": "@bayou/testing-common",
   "version": "1.0.0",
 
   "dependencies": {
