+++ local-modules/testing-client/package.json
@@ -7,7 +7,7 @@
     "@bayou/see-all": "local",
     "@bayou/promise-util": "local",
     "mocha-client-shim": "local",
-    "testing-common": "local",
+    "@bayou/testing-common": "local",
     "@bayou/util-common": "local"
   }
 }
