+++ local-modules/testing-server/CollectingReporter.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { BaseReporter } from 'testing-common';
+import { BaseReporter } from '@bayou/testing-common';
 
 import TestCollector from './TestCollector';
